Add Admin

+++ Library Managment Project/DTOs/MemberDTOs/GetAllMembers.Response.cs
@@ -4,7 +4,7 @@ namespace Library_Managment_Project.DTOs.MemberDTOs
 {
     public class GetAllMembersResponse
     {
-        public long MemberNumber { get; set; }
+        public int MemberNumber { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
@@ -13,7 +13,7 @@ namespace Library_Managment_Project.DTOs.MemberDTOs
         public List<string>? Bookloaned { get; set; }
         public DateTime CreateAt { get; set; }
         public DateTime UpdateAt { get; set; }
-        public GetAllMembersResponse (long memberNumber, string firstName, string lastName, string email, int phoneNumber, MemberShipType memberShipType, List<string> bookLoaned,DateTime createDate, DateTime updateDate)
+        public GetAllMembersResponse (int memberNumber, string firstName, string lastName, string email, int phoneNumber, MemberShipType memberShipType, List<string> bookLoaned,DateTime createDate, DateTime updateDate)
         {
             MemberNumber = memberNumber;
             FirstName = firstName;
