+++ Library Managment Project/DTOs/MemberDTOs/UpdateMember.Request.cs
@@ -4,12 +4,12 @@ namespace Library_Managment_Project.DTOs.MemberDTOs
 {
     public class UpdateMemberRequest
     {
-        public long MemberNumber { get; set; }
+        public int MemberNumber { get; set; }
         public string Email { get; set; }
         public int PhoneNumber { get; set; }
         public MemberShipType MemberShipType { get; set; }
         public DateTime UpdateAt { get; set; }
-        public UpdateMemberRequest(long memberNumber, string email, int phoneNumber, MemberShipType memberShipType, DateTime updateDate)
+        public UpdateMemberRequest(int memberNumber, string email, int phoneNumber, MemberShipType memberShipType, DateTime updateDate)
         {
 
             MemberNumber = memberNumber;
